third_party/blink/web_tests/editing/text-iterator/basic-iteration-shadowdom-expected.txt

Unit tests for WebCore text iterator with shadow tree support enabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS addShadowTreeWithDivElement(div); range.selectNodeContents(div.shadowRoot); internals.rangeAsText(range) is "b"
PASS appendBrElement(div.shadowRoot.childNodes[0]); range.selectNodeContents(div.shadowRoot); internals.rangeAsText(range) is "b\n"
PASS successfullyParsed is true

TEST COMPLETE
